+++ ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: misaev <misaev@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/31 10:29:03 by misaev            #+#    #+#             */
-/*   Updated: 2021/03/31 18:09:57 by misaev           ###   ########.fr       */
+/*   Updated: 2021/04/02 16:18:54 by misaev           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = ft_strlen(s);
 	write(fd, s, i);
