+++ ft_strmapi.c
@@ -6,16 +6,15 @@
 /*   By: misaev <misaev@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/31 09:03:16 by misaev            #+#    #+#             */
-/*   Updated: 2021/03/31 18:16:21 by misaev           ###   ########.fr       */
+/*   Updated: 2021/04/02 16:32:15 by misaev           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strmapi(char const *s, char (*f)(unsigned
-int, char))
+char	*ft_strmapi(char const *s, char (*f)(size_t, char))
 {
-	unsigned int	i;
+	size_t			i;
 	char			*str;
 
 	str = ft_strdup(s);
